+++ CutTheRope/Framework/Core/ResourceMgr.cs
@@ -291,7 +291,7 @@ namespace CutTheRope.Framework.Core
                 CTRRectangle rect = MakeRectangle(data[quadDataIndex], data[quadDataIndex + 1], data[quadDataIndex + 2], data[quadDataIndex + 3]);
                 if (lowestPoint < rect.h + rect.y)
                 {
-                    lowestPoint = (int)Ceil((double)(rect.h + rect.y));
+                    lowestPoint = (int)Ceil(rect.h + rect.y);
                 }
                 rect.x /= scaleX;
                 rect.y /= scaleY;
